pyproject.toml
[tool.poetry]
name = "build-an-agent"
version = "0.1.0"
description = "A simple demonstration of building AI agents using Salesforce Einstein LLM Gateway"
authors = ["Salesforce Engineering Team"]
readme = "README.md"
packages = [{include = "llms"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = ">=0.3.0"
langchain-core = ">=0.3.0"
langchain-openai = ">=0.2.0"
langgraph = ">=0.2.0"
pydantic = ">=2.0.0"
requests = ">=2.31.0"
python-dotenv = ">=1.0.0"
typing-extensions = ">=4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
weather-demo = "examples.weather_agent_demo:main" 